+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - @iov/socket: Add 10 second timeout to `SocketWrapper`. Value can be configured
   using an optional constructor argument.
-- @iov/socket: `StreamingSocket` now times out after 10 seconds.
+- @iov/socket: Add 10 second timeout to `StreamingSocket`. Value can be
+  configured using an optional constructor argument.
 
 ## 0.14.1
 
@@ -35,7 +36,8 @@
 
 - @iov/socket: Add 10 second timeout to `SocketWrapper`. Value can be configured
   using an optional constructor argument.
-- @iov/socket: `StreamingSocket` now times out after 10 seconds.
+- @iov/socket: Add 10 second timeout to `StreamingSocket`. Value can be
+  configured using an optional constructor argument.
 
 ## 0.13.6